Add files via upload
if all the brackets are placed correctly, the script will output SUCCESS.If the brackets are placed incorrectly, the editor should also inform the user of the first place where the error was detected. First of all, you need to find a closing bracket for which either there is no corresponding opening one (for example, the bracket ] in the string"] ()” ), or it closes an opening bracket that does not correspond to it (example: “()[}”). If there are no such errors, it is necessary to find the first opening bracket for which there is no corresponding closing one (example: parenthesis ( in the string" {} ([]”).
In addition to parentheses, the source code can contain Latin alphabet characters, numbers and punctuation marks.